Updated as per review comments
